refactor: 파일구조정리
