+++ src/main/java/org/example/ssj3pj/controller/youtube/ContentsController.java
@@ -1,4 +1,4 @@
-package org.example.ssj3pj.controller;
+package org.example.ssj3pj.controller.youtube;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
