+++ src/main/java/org/example/ssj3pj/controller/youtube/YoutubeAnalyticsController.java
@@ -1,11 +1,11 @@
-package org.example.ssj3pj.controller;
+package org.example.ssj3pj.controller.youtube;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.ssj3pj.dto.youtube.DailyDemographicsDto;
 import org.example.ssj3pj.dto.request.PeriodRequest;
 import org.example.ssj3pj.dto.youtube.TrafficSourceCategoryDto;
-import org.example.ssj3pj.services.YoutubeAnalyticsService;
+import org.example.ssj3pj.services.youtube.YoutubeAnalyticsService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
