+++ src/main/java/org/example/ssj3pj/services/youtube/YoutubeAnalyticsService.java
@@ -1,4 +1,4 @@
-package org.example.ssj3pj.services;
+package org.example.ssj3pj.services.youtube;
 
 import lombok.RequiredArgsConstructor;
 import org.example.ssj3pj.dto.youtube.DailyDemographicsDto;
